projecct/requirements.txt
Flask
Flask-WTF
requests
python-dotenv